+++ pagessite/src/components/utilities/swooch.svelte
@@ -1,0 +1,26 @@
+<!-- Created with Inkscape (http://www.inkscape.org/) -->
+<script lang="ts">
+	export let w: number;
+   export let h: number;
+   export let fill: string;
+</script>
+
+<!-- <?xml version="1.0" encoding="UTF-8" standalone="no"?> -->
+
+<svg
+	width={w}
+   height={h}
+	viewBox="0 0 181.73076 35.30769"
+	version="1.1"
+	id="svg5"
+	xmlns="http://www.w3.org/2000/svg"
+>
+	<g>
+		<path
+			style="fill:#000000;fill-opacity:1;stroke:none;stroke-width:0.265;stroke-linecap:butt;stroke-linejoin:miter;stroke-miterlimit:4;stroke-dasharray:none;stroke-opacity:1"
+			d="M 0,0 V 35.30769 L 181.73076,21.11538 12.115384,25.96154 Z"
+			id="swooch"
+         fill={fill}
+		/>
+	</g>
+</svg>
