+++ pagessite/src/data/b9902410d7136a4321df9443b4f7d461.json
@@ -127,7 +127,7 @@
                         "2": ""
                     },
                     "drops": [],
-                    "bye": false
+                    "bye": true
                 },
                 {
                     "p_ids": [
