+++ pagessite/src/data/bb2c0681137d3040162885512bfe23d9.json
@@ -72,7 +72,7 @@
                         "2": ""
                     },
                     "drops": [],
-                    "bye": false
+                    "bye": true
                 }
             ]
         },
